+++ NEWS.md
@@ -82,6 +82,7 @@ New library functions
   efficiently ([#35816]).
 * New function `addenv` for adding environment mappings into a `Cmd` object, returning the new `Cmd` object.
 * New function `insorted` for determining whether an element is in a sorted collection or not ([#37490]).
+* New function `Base.indent` for indenting non-blank lines by a number of spaces.
 
 New library features
 --------------------